requirements.txt
google-api-python-client
google-auth-oauthlib
feedparser
requests
google-generativeai